aac解码调整

+++ src/aacType.h
@@ -66,7 +66,7 @@ int aac_createHeader(AacHeader *head, uint8_t chn, uint16_t freq, uint16_t codeR
  *      bytesConsumed: 已使用in数据长度,用于in数据偏移,返回0时表示缺少数据量
  *  返回: pcm数据长度, -1/解析aac头失败, 0/数据不足,bytesConsumed返回缺少数据量
  */
-int aac_decode(void **aacDec, uint8_t *in, int inLen, uint8_t *out, int *bytesConsumed, int *chn, int *freq);
+int aac_decode(void **aacDec, uint8_t *in, int inLen, uint8_t *out, int *bytesConsumed, uint8_t *chn, uint16_t *freq);
 
 /*
  *  aac解码为pcm
@@ -76,7 +76,7 @@ int aac_decode(void **aacDec, uint8_t *in, int inLen, uint8_t *out, int *bytesCo
  *      out: 返回数据缓冲区,要求大于等于8192
  *  返回: pcm数据长度, -1失败
  */
-int aac_decode2(void **aacDec, int aacFile_fd, uint8_t *out, int *chn, int *freq);
+int aac_decode2(void **aacDec, int aacFile_fd, uint8_t *out, uint8_t *chn, uint16_t *freq);
 
 //文件
 void aac_decodeToFile(char *aacFile, char *pcmFile);
